main: move reload tests beside reload module
